backlight: qcom-wled: Add PMI8994 with different ovp/boost values.
PMI8994 has different curves to convert these values to indices written
to the appropriate registers. The values in the DT are matched exactly
to the values in the arrays, and a failed match fails the probe.
Especially ovp doesn't have a value close enough on the same index in
PMI8998 and up, which rules out changing the DT to abuse the wrong
values.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>